Updated the repo with corrections made to the code in the files

+++ src/Inventory/java/controller/ModifyProduct.java
@@ -63,6 +63,7 @@ public class ModifyProduct implements Initializable {
     /**
      * Returns the user to the main screen
      * @param event - The user is returned to the main menu
+     * @throws IOException
      */
     public void returnToMainMenu(ActionEvent event) throws IOException {
         Parent root = FXMLLoader.load(getClass().getResource("/view/Main.fxml"));
@@ -166,23 +167,52 @@ public class ModifyProduct implements Initializable {
      * @param event - Add is selected by the user
      */
     public void handleModProdAdd(ActionEvent event) {
-        Part item = modProdAssocPartsTable.getSelectionModel().getSelectedItem();
-        currProd.addAssociatedPart(item);
+        Part item = modProdPartsTable.getSelectionModel().getSelectedItem();
+        if(item != null){
+            currProd.addAssociatedPart(item);}
+        else{
+            Alert alert = new Alert(Alert.AlertType.ERROR);
+            alert.setTitle("ERROR");
+            alert.setHeaderText("Empty Selection");
+            alert.setContentText("You have not selected a part to add!");
+            alert.showAndWait();
+        }
     }
     /**
      * Takes the item that the user requests to remove, asks them if they are sure
      * they have selected the right item, and then removes the item.
      * @param event - Remove is clicked
+     *
+     *            <p> RUNTIME ERROR:
+     *  Encountered a NullPointerException error when I allowed the program to accept an item as null if the user
+     *  failed to select an item. The program attempted to retrieve the name of the item and that is where the error
+     *  occurred.
+     *  Implemented a try-catch block to circumvent this and check the user's selection, presenting them
+     *  with an error message if they had not selected an item.</p>
      */
     public void handleModProdRemove(ActionEvent event) {
         Part item = modProdAssocPartsTable.getSelectionModel().getSelectedItem();
-        Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
-        alert.setTitle("WARNING");
-        alert.setHeaderText("Part Removal Confirmation");
-        alert.setContentText("Are you sure you want to remove " + item.getName() + "?");
-        Optional<ButtonType> result = alert.showAndWait();
-        if (result.isPresent() && result.get() == ButtonType.OK) {
-            currProd.deleteAssociatedPart(item);
+        try {
+            if (item == null)
+                throw new Exception();
+
+            else {
+                Alert alert = new Alert(Alert.AlertType.CONFIRMATION);
+                alert.setTitle("WARNING");
+                alert.setHeaderText("Part Removal Confirmation");
+                alert.setContentText("Are you sure you want to remove " + item.getName() + "?");
+                Optional<ButtonType> result = alert.showAndWait();
+                if (result.isPresent() && result.get() == ButtonType.OK) {
+                    currProd.deleteAssociatedPart(item);
+                }
+            }
+        }
+        catch(Exception e){
+            Alert alert = new Alert(Alert.AlertType.ERROR);
+            alert.setTitle("ERROR");
+            alert.setHeaderText("Empty Selection");
+            alert.setContentText("You have not selected a part to remove!");
+            alert.showAndWait();
         }
     }
     /**
